fix(117): split callback state to separate file
